LPS-54985 Extract mappers and comparators
